Parse glsl files with windows carriage returns
Ran into an upstream failure in stackgl/headless-gl
when parsing preprocessor lines with windows carriage
returns (\r\n).

This adds a test case, and fixes it in two possible 
ways:
1. replace all \r with \n before parsing, or
2. on the preprocessor line, end at \r or \n.

The second fix is the smallest possible fix for the
issue I was running into. But the first fix seems like
a more general solution.